+++ src/core/index.ts
@@ -103,6 +103,7 @@ export type {
   TableProps,
   TableFilterProps,
   TableFilterValue,
+  DateRangeFilterOptions,
   FilterButtonBarProps,
 } from './Table';